pyproject.toml
[project]
name = "ratto"
version = "0.1.0"
description = "Your product wishlist"
authors = [
    {name = "capaths",email = "pathsankan@gmail.com"}
]
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (==5.2.1)",
    "djangorestframework (==3.16.0)",
    "markdown (==3.8)",
    "django-filter (==25.1)",
    "django-vite (==3.1.0)",
    "django-cors-headers (==4.7.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-nested-routers (>=0.94.2,<0.95.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "notebook (>=7.4.3,<8.0.0)",
    "selenium (>=4.33.0,<5.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "cloudscraper (>=1.2.71,<2.0.0)",
    "pytest (==8.3.5)"
]

[tool.poetry]
package-mode = false
